+++ packages/sdk/CHANGELOG.md
@@ -51,6 +51,7 @@
 
 ### Changed
 
+- Enable `denyList`/`allowList` validation on plt token transfers.
 - Client side validation in the PLT `Token` client is now disabled by default and has to be enabled explicitly if wanted
 
 ## 10.0.0-alpha.13